assets/Faktoren/requirements.txt
pandas
pocketbase
python-dotenv
openpyxl